main/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic-settings
sqlalchemy>=2.0
asyncpg
python-jose
passlib[bcrypt]
python-multipart